pyproject.toml
[project]
name = "s3-lambda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.221.0",
    "awswrangler>=3.14.0",
    "boto3>=1.41.2",
    "constructs>=10.4.2",
    "python-dotenv>=1.2.1",
]
